feat: broke out KeyValueRow stubs into their own directories
