+++ app/component-library/components-temp/KeyValueRow/KeyValueSection/KeyValueSection.styles.tsx
@@ -2,10 +2,8 @@ import { StyleSheet } from 'react-native';
 
 const styleSheet = () =>
   StyleSheet.create({
-    flexRow: {
-      flexDirection: 'row',
-      alignItems: 'center',
-      gap: 8,
+    keyValueSectionContainer: {
+      flex: 1,
     },
   });
